Migrate `options` test to `wit-bindgen test`
