--- crates/test-rust-wasm/src/bin/options.rs
@@ -1,3 +1,0 @@
-include!("../../../../tests/runtime/options/wasm.rs");
-
-fn main() {}
